# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the main API server for SciFIND, a scientific literature discovery platform. It provides endpoints for searching academic papers across multiple providers, managing papers and authors, and retrieving scientific literature metadata.",
        "title": "SciFIND Backend API",
        "termsOfService": "https://scifind.ai/terms",
        "contact": {
            "name": "SciFIND Support",
            "url": "https://scifind.ai/support",
            "email": "support@scifind.ai"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/search": {
            "get": {
                "description": "Search for academic papers across multiple providers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search for academic papers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to return (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of providers (arxiv,semantic_scholar,exa,tavily)",
                        "name": "providers",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date filter (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date filter (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Author filter",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Journal filter",
                        "name": "journal",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category filter",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scifind-backend_internal_services.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/search/papers/{provider}/{id}": {
            "get": {
                "description": "Retrieve a specific paper by provider name and paper ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Get a specific paper",
                "parameters": [
                    {
                        "enum": [
                            "arxiv",
                            "semantic_scholar",
                            "exa",
                            "tavily"
                        ],
                        "type": "string",
                        "description": "Provider name",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Paper ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scifind-backend_internal_services.PaperResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/search/providers": {
            "get": {
                "description": "Get information about available search providers and their status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Get available search providers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scifind-backend_internal_services.ProviderStatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/search/providers/{provider}/configure": {
            "put": {
                "description": "Update the configuration of a specific search provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Configure a search provider",
                "parameters": [
                    {
                        "enum": [
                            "arxiv",
                            "semantic_scholar",
                            "exa",
                            "tavily"
                        ],
                        "type": "string",
                        "description": "Provider name",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Provider configuration",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scifind-backend_internal_providers.ProviderConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scifind-backend_internal_services.ProviderConfigResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_api_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "scifind-backend_internal_models.Paper": {
            "type": "object",
            "required": [
                "authors",
                "id",
                "language",
                "source_id",
                "source_provider",
                "title"
            ],
            "properties": {
                "abstract": {
                    "type": "string",
                    "maxLength": 10000
                },
                "arxiv_id": {
                    "type": "string"
                },
                "authors": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/scifind-backend_internal_models.Author"
                    }
                },
                "categories": {
                    "description": "Classification and metrics",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/scifind-backend_internal_models.Category"
                    }
                },
                "citation_count": {
                    "description": "Citation metrics",
                    "type": "integer",
                    "minimum": 0
                },
                "citations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "description": "Timestamps",
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "doi": {
                    "type": "string"
                },
                "extracted_data": {
                    "type": "string"
                },
                "full_text": {
                    "description": "Content analysis",
                    "type": "string"
                },
                "id": {
                    "description": "Primary identifiers",
                    "type": "string"
                },
                "issue": {
                    "type": "string",
                    "maxLength": 50
                },
                "journal": {
                    "description": "Publication details",
                    "type": "string",
                    "maxLength": 500
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "language": {
                    "type": "string"
                },
                "pages": {
                    "type": "string",
                    "maxLength": 100
                },
                "pdf_url": {
                    "type": "string",
                    "maxLength": 2048
                },
                "processing_state": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "processing",
                        "completed",
                        "failed"
                    ]
                },
                "published_at": {
                    "type": "string"
                },
                "quality_score": {
                    "description": "Quality and processing",
                    "type": "number",
                    "maximum": 1,
                    "minimum": 0
                },
                "references": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "source_id": {
                    "type": "string"
                },
                "source_provider": {
                    "description": "Source tracking",
                    "type": "string",
                    "enum": [
                        "arxiv",
                        "semantic_scholar",
                        "exa",
                        "tavily",
                        "manual"
                    ]
                },
                "source_url": {
                    "type": "string"
                },
                "title": {
                    "description": "Core metadata",
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "description": "URLs and access",
                    "type": "string",
                    "maxLength": 2048
                },
                "volume": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
        "scifind-backend_internal_providers.CacheConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "key_prefix": {
                    "type": "string"
                },
                "max_size": {
                    "type": "integer"
                },
                "ttl": {
                    "$ref": "#/definitions/time.Duration"
                }
            }
        },
        "scifind-backend_internal_providers.CircuitBreakerConfig": {
            "type": "object",
            "properties": {
                "failure_threshold": {
                    "type": "integer"
                },
                "interval": {
                    "$ref": "#/definitions/time.Duration"
                },
                "max_requests": {
                    "type": "integer"
                },
                "success_threshold": {
                    "type": "integer"
                },
                "timeout": {
                    "$ref": "#/definitions/time.Duration"
                }
            }
        },
        "scifind-backend_internal_providers.ProviderConfig": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "api_secret": {
                    "type": "string"
                },
                "base_url": {
                    "description": "API configuration",
                    "type": "string"
                },
                "cache": {
                    "description": "Caching",
                    "allOf": [
                        {
                            "$ref": "#/definitions/scifind-backend_internal_providers.CacheConfig"
                        }
                    ]
                },
                "circuit_breaker": {
                    "description": "Circuit breaker",
                    "allOf": [
                        {
                            "$ref": "#/definitions/scifind-backend_internal_providers.CircuitBreakerConfig"
                        }
                    ]
                },
                "custom": {
                    "description": "Custom settings",
                    "type": "object",
                    "additionalProperties": true
                },
                "enabled": {
                    "type": "boolean"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "max_retries": {
                    "type": "integer"
                },
                "name": {
                    "description": "Basic config",
                    "type": "string"
                },
                "rate_limit": {
                    "description": "Rate limiting",
                    "allOf": [
                        {
                            "$ref": "#/definitions/scifind-backend_internal_providers.RateLimitConfig"
                        }
                    ]
                },
                "retry_delay": {
                    "$ref": "#/definitions/time.Duration"
                },
                "timeout": {
                    "description": "Timeouts",
                    "allOf": [
                        {
                            "$ref": "#/definitions/time.Duration"
                        }
                    ]
                }
            }
        },
        "scifind-backend_internal_providers.ProviderError": {
            "type": "object",
            "properties": {
                "error": {},
                "provider": {
                    "type": "string"
                },
                "retryable": {
                    "type": "boolean"
                },
                "type": {
                    "description": "timeout, rate_limit, network, parse, etc.",
                    "type": "string"
                }
            }
        },
...
        "scifind-backend_internal_services.PaperResponse": {
            "type": "object",
            "properties": {
                "paper": {
                    "$ref": "#/definitions/scifind-backend_internal_models.Paper"
                },
                "source": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "scifind-backend_internal_services.ProviderConfigResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "provider_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/scifind-backend_internal_providers.ProviderStatus"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "scifind-backend_internal_services.ProviderStatusResponse": {
            "type": "object",
            "properties": {
                "providers": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/scifind-backend_internal_providers.ProviderStatus"
                    }
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "scifind-backend_internal_services.SearchResponse": {
            "type": "object",
            "properties": {
                "aggregation_strategy": {
                    "type": "string"
                },
                "cache_hits": {
                    "type": "integer"
                },
                "duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/scifind-backend_internal_providers.ProviderError"
                    }
                },
                "papers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/scifind-backend_internal_models.Paper"
                    }
                },
                "partial_failure": {
                    "type": "boolean"
                },
                "providers_failed": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "providers_used": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "query": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "result_count": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "time.Duration": {
            "type": "integer",
            "format": "int64",
            "enum": [
                -9223372036854775808,
                9223372036854775807,
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000
            ],
            "x-enum-varnames": [
                "minDuration",
                "maxDuration",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour"
            ]
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}